Update index.html
the prompt i gave chatgpt: what I want to do now is change the styling on the index.html code you gave me. I would not like a contact or about section. I would love if instead the design could have the text be in a white square in the middle with a purple background. it would be great if the button could be a white box with a purple arrow. can you generate me that code? I still want the button to go to my other file btw (its titled questions2.html)